improvements to UI
- factorized languages/tags/offliners/defs into single requests at start
- factorized schedule retrieval for view/edit
- expiring cookie after 1d
- better signoff message (duration)
- moved editor code to custom component
- revamped edition detection/payload
