+++ dispatcher/frontend-ui/package.json
@@ -19,6 +19,7 @@
     "filesize": "^6.0.1",
     "moment": "^2.24.0",
     "querystring": "^0.2.0",
+    "sugar": "^2.0.6",
     "vue": "^2.6.10",
     "vue-clipboard2": "^0.3.1",
     "vue-cookie": "^1.1.4",
