+++ dispatcher/frontend-ui/src/components/Mixins.js
@@ -9,9 +9,12 @@ export default {
      };
   },
   computed: {
-    publicPath() {
-      return process.env.BASE_URL;
-    },
+    publicPath() { return process.env.BASE_URL; },  // for static files linking
+    categories() { return Constants.categories; }, // list of categories for filter/edit
+    warehouse_paths() { return Constants.warehouse_paths; },  // list of paths for edit
+    languages() { return this.$store.getters.languages; },  // list of lang for filter/edit
+    tags() { return this.$store.getters.tags; },  // list of tags for filter/edit
+    offliners() { return this.$store.getters.offliners; }, // list of offliners for edit
   },
   methods: {
     scrollToTop() { window.scrollTo(0,0); },
@@ -20,10 +23,10 @@ export default {
       let payload = text ? {status: true, text: text} : {status: false};
       this.$store.dispatch('setLoading', payload);
     },
-    getLimitPreference() {
+    getLimitPreference() {   // retrieve nb of items to display from cookie
       return this.$cookie.get('pref-limit') || Constants.DEFAULT_LIMIT;
     },
-    saveLimitPreference(value) {
+    saveLimitPreference(value) {  // save nb of items to display in a cookie
       this.$cookie.set('pref-limit', value);
     },
   }
